Refactor data module
