+++ texar/torch/data/data/__init__.py
@@ -22,5 +22,6 @@ from texar.torch.data.data.mono_text_data import *
 from texar.torch.data.data.multi_aligned_data import *
 from texar.torch.data.data.paired_text_data import *
 from texar.torch.data.data.record_data import *
+from texar.torch.data.data.sampler import *
 from texar.torch.data.data.scalar_data import *
 from texar.torch.data.data.text_data_base import *
