+++ texar/torch/data/data/data_iterators.py
@@ -18,17 +18,18 @@ Various data iterator classes.
 # pylint: disable=protected-access
 
 from typing import (
-    Any, Callable, Dict, Generic, Iterable, Iterator, List, Optional,
-    Sequence, Tuple, TypeVar, Union, Mapping)
+    Dict, Iterable, Iterator, List, Optional, Sequence, Union, Mapping)
 
 import pkg_resources
 import torch
 from torch import __version__ as _torch_version  # type: ignore
 from torch.utils.data import DataLoader
-from torch.utils.data import sampler as torch_sampler
 
 from texar.torch.data.data.data_base import DatasetBase
 from texar.torch.data.data.dataset_utils import Batch
+from texar.torch.data.data.sampler import (
+    SamplerBase, SequentialSampler, RandomSampler, BufferShuffleSampler,
+    BatchingStrategy, DynamicBatchSampler)
 from texar.torch.utils.types import MaybeSeq
 from texar.torch.utils.utils import ceildiv, map_structure
 
@@ -37,315 +38,10 @@ _torch_version = pkg_resources.parse_version(_torch_version)
 __all__ = [
     "DataIterator",
     "TrainTestDataIterator",
-    "BatchingStrategy",
-    "TokenCountBatchingStrategy",
 ]
 
 # `Dict` is invariant, `Mapping` is not.
 DatasetsType = Union[Mapping[str, DatasetBase], MaybeSeq[DatasetBase]]
-Example = TypeVar('Example')
-
-
-# pylint: disable=attribute-defined-outside-init
-# TODO: Remove this when Pylint fixes the bug. If the `disable` directive is not
-#  added, Pylint incorrectly reports this error for `self.size` in subclasses of
-#  `SamplerBase` in Python 3.6 due to use of the Generic class.
-#  See Pylint issue: https://github.com/PyCQA/pylint/issues/2981
-
-class SamplerBase(torch_sampler.Sampler, Generic[Example]):
-    r"""A subclass of :torch_docs:`~torch.utils.data.Sampler
-    <data.html#torch.utils.data.Sampler>` that supports:
-
-    - Returning raw examples when required.
-    - Creating iterators with unknown dataset size.
-
-    This class is used internally in
-    :class:`~texar.torch.data.data.DataIterator`. It calls the
-    :meth:`~texar.torch.data.data.DatasetBase._prefetch_source` method to ensure
-    the required number of raw examples are prefetched from source.
-
-    Args:
-        data: The :class:`~texar.torch.data.data.DatasetBase` instance.
-    """
-    size: Optional[int]
-
-    def __init__(self, data: DatasetBase[Any, Example]):
-        super().__init__(data)
-
-        self._data = data
-        self.size = None
-
-    def _iterator_given_size(self, size: int) -> Iterator[int]:
-        r"""Return an iterator that generates samples when the dataset size
-        is given.
-
-        Args:
-            size: The dataset size.
-        """
-        raise NotImplementedError
-
-    def _iterator_unknown_size(self) -> Iterator[int]:
-        r"""Return an iterator that generates samples when the dataset size
-        is unknown. This iterator must also call
-        :meth:`texar.torch.data.data.DatasetBase._prefetch_source` and check
-        whether the dataset size can be determined, before yielding the index.
-        See example implementations for details.
-        """
-        raise NotImplementedError
-
-    def __iter__(self) -> Union[Iterator[int], Iterator[Tuple[int, Example]]]:
-        r"""Return an iterator based on the dataset settings.
-        """
-        self.size = self._data._dataset_size
-        if (not self._data._fully_cached or
-                self._data._should_call_prefetch_source):
-            self._data._start_iteration()
-            # First epoch of lazy loading, calling prefetch, and returning
-            # indices and examples.
-            iterator = self._iterator_unknown_size()
-        else:
-            # Non-lazy loading, or when dataset has been fully iterated.
-            assert self.size is not None
-            iterator = self._iterator_given_size(self.size)
-
-        if self._data._should_call_prefetch_processed:
-            # Processing routine is performed in main process. Yield
-            # processed examples instead.
-            map_fn = lambda idx: (idx, self._data._processed_cache[idx])
-        elif self._data._should_yield_raw_example:
-            # Return indices and examples for any epoch in this case.
-            map_fn = lambda idx: (idx, self._data._source[idx])
-        else:
-            map_fn = None  # type: ignore
-        if map_fn is not None:
-            return map(map_fn, iterator)
-
-        return iterator
-
-    def __len__(self):
-        if self.size is not None:
-            return self.size
-        raise AttributeError("Dataset size cannot be determined at this point")
-
-
-class SequentialSampler(SamplerBase[Example]):
-    r"""Samples elements sequentially, always in the same order. Same as
-    :torch_docs:`~torch.utils.data.SequentialSampler
-    <data.html#torch.utils.data.SequentialSampler>`
-    """
-
-    def _iterator_given_size(self, size: int) -> Iterator[int]:
-        return iter(range(size))
-
-    def _iterator_unknown_size(self) -> Iterator[int]:
-        index = 0
-        while True:
-            cur_size = self._data._prefetch_source(index)
-            if cur_size is not None:
-                self.size = cur_size
-                break
-            yield index
-            index += 1
-
-
-class RandomSampler(SamplerBase[Example]):
-    r"""Samples elements randomly. If without replacement, then sample from a
-    shuffled dataset. If with replacement, then user can specify ``num_samples``
-    to draw.
-
-    This class uses :torch_docs:`torch.utils.data.RandomSampler
-    <data.html#torch.utils.data.RandomSampler>` directly. Given the
-    nature of such shuffling, it cannot be used for iterators with unknown size.
-
-    Args:
-        data: The :class:`~texar.torch.data.data.DatasetBase` instance.
-        num_samples (int): number of samples to draw, default=len(dataset)
-        replacement (bool): samples are drawn with replacement if `True`,
-            default=False
-    """
-
-    def __init__(self, data: DatasetBase[Any, Example],
-                 replacement: bool = False, num_samples: Optional[int] = None):
-        super().__init__(data)
-        self._sampler = torch_sampler.RandomSampler(
-            data, replacement, num_samples)
-
-    def _iterator_given_size(self, size: int) -> Iterator[int]:
-        del size  # not used
-        return iter(self._sampler)
-
-    def _iterator_unknown_size(self) -> Iterator[int]:
-        raise TypeError(
-            "RandomSampler does not support lazy data loading. To perform "
-            "shuffling with lazy loading, use BufferShuffleSampler.")
-
-
-class BufferShuffleSampler(SamplerBase[Example]):
-    r"""A :torch_docs:`~torch.utils.data.Sampler
-    <data.html#torch.utils.data.Sampler>` that uses a shuffle buffer, as
-    in TensorFlow. The buffer is first filled with data examples. Each time a
-    sample is drawn from the buffer, and the drawn sample is replaced with the
-    next data example.
-
-    This class is used internally in
-    :class:`~texar.torch.data.data.DataIterator`. It calls the
-    :meth:`~texar.torch.data.data.DatasetBase._prefetch_source` method to ensure
-    the required number of raw examples are prefetched from source.
-
-    Args:
-        data: The :class:`~texar.torch.data.data.DatasetBase` instance.
-        buffer_size: The size of the shuffle buffer. Use larger buffer sizes for
-            more uniformly-random shuffling.
-    """
-
-    def __init__(self, data: DatasetBase[Any, Example], buffer_size: int):
-        super().__init__(data)
-        self.buffer_size = buffer_size
-
-    def _iterator_given_size(self, size) -> Iterator[int]:
-        if self.buffer_size >= size:
-            yield from iter(torch.randperm(size).tolist())
-            return
-
-        buffer = list(range(self.buffer_size))
-        for x in range(self.buffer_size, size):
-            sample = torch.randint(self.buffer_size, (1,)).item()
-            index = buffer[sample]
-            yield index
-            buffer[sample] = x
-        yield from (buffer[x] for x in torch.randperm(self.buffer_size))
-
-    def _iterator_unknown_size(self) -> Iterator[int]:
-        buffer = list(range(self.buffer_size))
-        x = self.buffer_size
-        while True:
-            sample = torch.randint(self.buffer_size, (1,)).item()
-            index = buffer[sample]
-            cur_size = self._data._prefetch_source(index)
-            if cur_size is not None and index >= cur_size:
-                self.size = cur_size
-            if self.size is not None and index >= self.size:
-                break
-            yield index
-            buffer[sample] = x
-            x += 1
-        yield from (buffer[x] for x in torch.randperm(self.buffer_size)
-                    if buffer[x] < self.size)
-
-
-# pylint: enable=attribute-defined-outside-init
-
-
-class BatchingStrategy(Generic[Example]):
-    r"""Decides batch boundaries in dynamic batching. Please refer to
-    :class:`TokenCountBatchingStrategy` for a concrete example.
-    """
-
-    def reset_batch(self) -> None:
-        r"""Reset the internal state of the batching strategy. This method is
-        called at the start of iteration, and after each batch is yielded.
-        """
-        raise NotImplementedError
-
-    def add_example(self, example: Example) -> bool:
-        r"""Add an example into the current batch, and modify internal states
-        accordingly. If the example should not be added to the batch, this
-        method does not modify the internal state, and returns `False`.
-
-        Args:
-            example: The example to add to the batch.
-
-        Returns:
-            A boolean value indicating whether :attr:`example` should be added
-            to the batch.
-        """
-        raise NotImplementedError
-
-
-class TokenCountBatchingStrategy(BatchingStrategy[Example]):
-    r"""Create dynamically-sized batches so that the total number of tokens
-    inside each batch is constrained.
-
-    Args:
-        max_tokens (int): The maximum number of tokens inside each batch.
-        max_batch_size (int, optional): The maximum number of examples for each
-            batch. If `None`, batches can contain arbitrary number of examples
-            as long as the total number of tokens does not exceed
-            :attr:`max_tokens`.
-        length_fn (callable, optional): A function taking a data example as
-            argument, and returning the number of tokens in the example. By
-            default, :python:`len` is used, which is the desired behavior if the
-            dataset in question is a :class:`~texar.torch.data.MonoTextData`.
-    """
-    sum_tokens: int
-    cur_batch_size: int
-
-    def __init__(self, max_tokens: int, max_batch_size: Optional[int] = None,
-                 length_fn: Optional[Callable[[Example], int]] = None):
-        self.max_batch_size = max_batch_size
-        self.max_tokens = max_tokens
-        self.length_fn: Callable[[Example], int]
-        self.length_fn = length_fn or len  # type: ignore
-
-    def reset_batch(self) -> None:
-        self.sum_tokens = 0
-        self.cur_batch_size = 0
-
-    def add_example(self, example: Example) -> bool:
-        if self.cur_batch_size == self.max_batch_size:
-            return False
-        cur_tokens = self.length_fn(example)
-        if cur_tokens + self.sum_tokens > self.max_tokens:
-            return False
-
-        self.cur_batch_size += 1
-        self.sum_tokens += cur_tokens
-        return True
-
-
-class DynamicBatchSampler(torch_sampler.BatchSampler, Generic[Example]):
-    r"""A subclass of :torch_docs:`~torch.utils.data.BatchSampler
-    <data.html#torch.utils.data.BatchSampler>` that supports dynamic batching
-    through a user-provided :class:`BatchingStrategy`. This class is used
-    internally.
-
-    Args:
-        dataset: The dataset to create batches from.
-        sampler: An instance of :class:`SamplerBase` that returns indices of
-            each sampled example.
-        strategy: An instance of :class:`BatchingStrategy` that decides whether
-            a batch should be yielded.
-    """
-
-    def __init__(self, dataset: DatasetBase[Any, Example],  # pylint: disable=super-init-not-called
-                 sampler: SamplerBase, strategy: BatchingStrategy[Example]):
-        self.dataset = dataset
-        self.sampler = sampler
-        self.strategy = strategy
-
-    def __iter__(self) -> Union[Iterator[List[int]],  # type: ignore
-                                Iterator[List[Tuple[int, Example]]]]:
-        batch = []  # type: ignore
-        self.strategy.reset_batch()
-        for idx in self.sampler:
-            if isinstance(idx, tuple):
-                example = self.dataset[idx[0]]
-            else:
-                example = self.dataset[idx]
-            while not self.strategy.add_example(example):
-                if len(batch) == 0:
-                    raise ValueError(f"Batching strategy refused to add "
-                                     f"example {idx} to empty batch.")
-                yield batch
-                batch = []
-                self.strategy.reset_batch()
-            batch.append(idx)
-        if len(batch) > 0:
-            yield batch
-            self.strategy.reset_batch()
-
-    def __len__(self):
-        raise TypeError("DynamicBatchSampler does not support __len__")
 
 
 # pylint: disable=ungrouped-imports
@@ -391,22 +87,13 @@ if _torch_version >= pkg_resources.parse_version("1.2.0"):  # PyTorch 1.2.0 +
     # classes, and construct concrete classes in their `__new__` methods
     # depending on the value of `num_workers`. This is for compatibility with
     # previous versions, so we don't need to change other parts of the code.
-    #
-    # PyTorch 1.3 change some attribute names in `_BaseDataLoaderIter`
-    # listed as follows:
-    # self.dataset ---> self._dataset
-    # self.dataset_fetcher ---> self._dataset_fetcher
-    # self.pin_memory --- > self._pin_memory
-    # self.tasks_outstanding ---> self._tasks_outstanding
-    # self.num_workers ---> self._num_workers
-    # self.index_queues ---> self._index_queues
-    # self.send_idx ---> self._send_idx
-    # self.task_info ---> self._task_info
-    # self.rcvd_idx ---> self._rcvd_idx
 
-    from torch.utils.data.dataloader import (  # type: ignore
-        _BaseDataLoaderIter, _SingleProcessDataLoaderIter,
-        _MultiProcessingDataLoaderIter)
+    from texar.torch.data.data.data_iterators_utils import \
+        TexarBaseDataLoaderIter as _BaseDataLoaderIter
+    from texar.torch.data.data.data_iterators_utils import \
+        TexarSingleProcessDataLoaderIter as _SingleProcessDataLoaderIter
+    from texar.torch.data.data.data_iterators_utils import \
+        TexarMultiProcessingDataLoaderIter as _MultiProcessingDataLoaderIter
 
     class _DataLoaderIter(_BaseDataLoaderIter):
         r"""Iterates once over the DataLoader's dataset. This is almost
@@ -433,16 +120,10 @@ if _torch_version >= pkg_resources.parse_version("1.2.0"):  # PyTorch 1.2.0 +
             batch = super().__next__()
             # Drop smaller final batch according to settings. Note that
             # `_batch_size` could be None if dynamic batching is used.
-            if _torch_version >= pkg_resources.parse_version("1.3.0"):
-                if (self._batch_size is not None and
-                        batch.batch_size < self._batch_size and
-                        not self._dataset.hparams.allow_smaller_final_batch):
-                    raise StopIteration
-            else:
-                if (self._batch_size is not None and
-                        batch.batch_size < self._batch_size and
-                        not self.dataset.hparams.allow_smaller_final_batch):
-                    raise StopIteration
+            if (self._batch_size is not None and
+                    batch.batch_size < self._batch_size and
+                    not self.dataset.hparams.allow_smaller_final_batch):
+                raise StopIteration
             if self.device is not None:
                 batch = move_memory(batch, self.device)
             return batch
@@ -478,24 +159,13 @@ if _torch_version >= pkg_resources.parse_version("1.2.0"):  # PyTorch 1.2.0 +
                                  _SingleProcessDataLoaderIter):
         def __next__(self):
             index = self._next_index()  # may raise StopIteration
-            if _torch_version >= pkg_resources.parse_version("1.3.0"):
-                # may raise StopIteration
-                data = self._dataset_fetcher.fetch(index)
-                if self._dataset._should_yield_raw_example:
-                    index = [idx[0] for idx in index]
-                examples, data = data
-                self._dataset._add_cached_examples(index, examples)
-                if self._pin_memory:
-                    data = move_memory(_pin_memory(data), self.device)
-            else:
-                # may raise StopIteration
-                data = self.dataset_fetcher.fetch(index)
-                if self.dataset._should_yield_raw_example:
-                    index = [idx[0] for idx in index]
-                examples, data = data
-                self.dataset._add_cached_examples(index, examples)
-                if self.pin_memory:
-                    data = move_memory(_pin_memory(data), self.device)
+            data = self.dataset_fetcher.fetch(index)  # may raise StopIteration
+            if self.dataset._should_yield_raw_example:
+                index = [idx[0] for idx in index]
+            examples, data = data
+            self.dataset._add_cached_examples(index, examples)
+            if self.pin_memory:
+                data = move_memory(_pin_memory(data), self.device)
             return data
 
     class _MPCacheDataLoaderIter(_CacheDataLoaderIter,
@@ -505,78 +175,41 @@ if _torch_version >= pkg_resources.parse_version("1.2.0"):  # PyTorch 1.2.0 +
         worker_queue_idx: int  # so that Pylint gives no errors
 
         def _try_put_index(self):
-            if _torch_version >= pkg_resources.parse_version("1.3.0"):
-                assert self._tasks_outstanding < 2 * self._num_workers
-                try:
-                    index = self._next_index()
-                except StopIteration:
-                    return
-                # find next active worker, if any
-                for _ in range(self._num_workers):
-                    worker_queue_idx = next(self._worker_queue_idx_cycle)
-                    if self._workers_status[worker_queue_idx]:
-                        break
-                else:
-                    # not found (i.e., didn't break)
-                    return
-
-                self._index_queues[worker_queue_idx].put((self._send_idx,
-                                                          index))
-                if self._dataset._should_yield_raw_example:
-                    index = [idx[0] for idx in index]
-                self._indices_dict[self._send_idx] = index
-                self._task_info[self._send_idx] = (worker_queue_idx,)
-                self._tasks_outstanding += 1
-                self._send_idx += 1
+            assert self.tasks_outstanding < 2 * self.num_workers
+            try:
+                index = self._next_index()
+            except StopIteration:
+                return
+            for _ in range(self.num_workers):  # find next active worker, if any
+                worker_queue_idx = next(self.worker_queue_idx_cycle)
+                if self.workers_status[worker_queue_idx]:
+                    break
             else:
-                assert self.tasks_outstanding < 2 * self.num_workers
-                try:
-                    index = self._next_index()
-                except StopIteration:
-                    return
-                # find next active worker, if any
-                for _ in range(self.num_workers):
-                    worker_queue_idx = next(self.worker_queue_idx_cycle)
-                    if self.workers_status[worker_queue_idx]:
-                        break
-                else:
-                    # not found (i.e., didn't break)
-                    return
+                # not found (i.e., didn't break)
+                return
 
-                self.index_queues[worker_queue_idx].put((self.send_idx, index))
-                if self.dataset._should_yield_raw_example:
-                    index = [idx[0] for idx in index]
-                self._indices_dict[self.send_idx] = index
-                self.task_info[self.send_idx] = (worker_queue_idx,)
-                self.tasks_outstanding += 1
-                self.send_idx += 1
+            self.index_queues[worker_queue_idx].put((self.send_idx, index))
+            if self.dataset._should_yield_raw_example:
+                index = [idx[0] for idx in index]
+            self._indices_dict[self.send_idx] = index
+            self.task_info[self.send_idx] = (worker_queue_idx,)
+            self.tasks_outstanding += 1
+            self.send_idx += 1
 
         def _process_data(self, batch):
             batch = super()._process_data(batch)
-            if _torch_version >= pkg_resources.parse_version("1.3.0"):
-                indices = self._indices_dict[self._rcvd_idx - 1]
-                del self._indices_dict[self._rcvd_idx - 1]
-                examples, batch = batch
-                self._dataset._add_cached_examples(indices, examples)
-            else:
-                indices = self._indices_dict[self.rcvd_idx - 1]
-                del self._indices_dict[self.rcvd_idx - 1]
-                examples, batch = batch
-                self.dataset._add_cached_examples(indices, examples)
+            indices = self._indices_dict[self.rcvd_idx - 1]
+            del self._indices_dict[self.rcvd_idx - 1]
+            examples, batch = batch
+            self.dataset._add_cached_examples(indices, examples)
             return batch
 
         def __next__(self):
             batch = super().__next__()
-            if _torch_version >= pkg_resources.parse_version("1.3.0"):
-                if (self._batch_size is not None and
-                        batch.batch_size < self._dataset.batch_size and
-                        not self._dataset.hparams.allow_smaller_final_batch):
-                    raise StopIteration
-            else:
-                if (self._batch_size is not None and
-                        batch.batch_size < self.dataset.batch_size and
-                        not self.dataset.hparams.allow_smaller_final_batch):
-                    raise StopIteration
+            if (self._batch_size is not None and
+                    batch.batch_size < self.dataset.batch_size and
+                    not self.dataset.hparams.allow_smaller_final_batch):
+                raise StopIteration
             batch = move_memory(batch, self.device)
             return batch
 else:
